# servidor/licenca/src/main/java/br/com/licenca/entity/response/LicencaResponse.java
package br.com.licenca.entity.response;

import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.Setter;

@Getter
@Setter
@AllArgsConstructor
public class LicencaResponse {
    private String valor;
}