# servidor/licenca/src/main/resources/static/swagger-apis/api/swagger.yaml
swagger: "2.0"
info:
    version: "1.0.0"
    title: "Servidor de Licenças de Software"
    license:
        name: "Apache 2.0"
basePath: "/v2"
tags:
  - name: "Licença"
    description: "licença"
schemes:
- "http"
paths:
  /licenca:
    get:
        tags:
        - "Licença"
        summary: "acquire"
        description: "Solicita uma licença para uso durante 1 minuto"
        operationId: "obterLicenca"
        consumes:
        - "application/json"
        produces:
        - "application/json"
        parameters:
          - name: "usuarioId"
            in: "query"
            description: "Id do usuário"
            required: true
            type: "string"
        responses:
            "200":
                description: Licença retornada com sucesso
                schema:
                    $ref: "#/definitions/LicencaResponse"
            "412":
                description: Todas as licenças disponíveis estão sendo utilizadas, Usuario informado não existe
                schema:
                    $ref: "#/definitions/ExceptionSchema"
                        
  /licenca/renovar:
    patch:
        tags:
        - "Licença"
        summary: "renew"
        description: "Renova uma solicitação de licença por mais 1 minuto"
        operationId: "renovarLicenca"
        consumes:
        - "application/json"
        produces:
        - "application/json"
        parameters:
          - name: "usuarioId"
            in: "query"
            description: "Id do usuário"
            required: true
            type: "string"
          - name: "licencaCodigo"
            in: "query"
            description: "Código da licença"
            required: true
            type: "string"  
        responses:
            "204":
                description: Licença renovada com sucesso
            "412":
                description: Usuário não existe, licença não existe, licença não está associada ao usuário
                schema:
                    $ref: "#/definitions/ExceptionSchema"
                
  /licenca/devolver:
    delete:
        tags:
        - "Licença"
        summary: "release"
        description: "Devolve a licença ao servidor após sua utilização"
        operationId: "devolverLicenca"
        consumes:
        - "application/json"
        produces:
        - "application/json"
        parameters:
          - name: "usuarioId"
            in: "query"
            description: "Id do usuário"
            required: true
            type: "string"
          - name: "licencaCodigo"
            in: "query"
            description: "Código da licença"
            required: true
            type: "string" 
        responses:
            "204":
                description: Licença devolvida com sucesso
            "412":
                description: Usuário não existe, licença não existe, licença não está associada ao usuário
                schema:
                    $ref: "#/definitions/ExceptionSchema"
                          
definitions: 
  LicencaResponse:
    type: "object"
    properties:
        valor:
            type: string
            example: "DE04D58DC5CCC4B9671C3627FB8D626FE4A15810BC1FE3E724FEEA761965FB71"
  
  ExceptionSchema:
      type: "object"
      properties:
          mensagem:
            type: string
            example: "Não há licença disponível no momento"
          timestamp:
            type: string
            example: "2022-03-12T20:45:16.226+0000"
          status: 
            type: integer
            example: 412,
          error: 
            type: string
            example: "Precondition Failed"
          message: 
            type: string
            example: "Usuário 2321 não existe"
          path: 
            type: string
            example: "/licenca"